+++ sources/niUI/src/GDRV_GLContext.cpp
@@ -137,9 +137,9 @@ int tsglCreateContext(
         WGL_SUPPORT_OPENGL_ARB, GL_TRUE,
         WGL_DOUBLE_BUFFER_ARB, GL_TRUE,
         WGL_PIXEL_TYPE_ARB, WGL_TYPE_RGBA_ARB,
-        WGL_COLOR_BITS_ARB, anColorBits,
-        WGL_DEPTH_BITS_ARB, anDepthBits,
-        WGL_STENCIL_BITS_ARB, anStencilBits,
+        WGL_COLOR_BITS_ARB, (int)anColorBits,
+        WGL_DEPTH_BITS_ARB, (int)anDepthBits,
+        WGL_STENCIL_BITS_ARB, (int)anStencilBits,
         WGL_SAMPLE_BUFFERS_ARB, 1, //Number of buffers (must be 1 at time of writing)
         WGL_SAMPLES_ARB, 4,        //Number of samples
         0
@@ -162,9 +162,9 @@ int tsglCreateContext(
         WGL_PIXEL_TYPE_ARB, WGL_TYPE_RGBA_ARB,
         WGL_DRAW_TO_WINDOW_ARB, true,
         WGL_DOUBLE_BUFFER_ARB, true,
-        WGL_COLOR_BITS_ARB, anColorBits,
-        WGL_DEPTH_BITS_ARB, anDepthBits,
-        WGL_STENCIL_BITS_ARB, anStencilBits,
+        WGL_COLOR_BITS_ARB, (int)anColorBits,
+        WGL_DEPTH_BITS_ARB, (int)anDepthBits,
+        WGL_STENCIL_BITS_ARB, (int)anStencilBits,
         0
       };
